Some more codes to handle authorization

+++ Core/MiniETrade.Domain/Entities/Product.cs
@@ -11,7 +11,7 @@ public class Product :BaseEntity
 {
     public string Name { get; set; }
     public int Stock { get; set; }
-    public float Price { get; set; }
+    public decimal Price { get; set; }
 
     public virtual ICollection<Order>? Orders { get; set; }
     public virtual ICollection<ProductImageFile>? ProductImages { get; set; }
